+++ src/rpclib/test/interop/test_client_zeromq.py
@@ -20,12 +20,12 @@
 import unittest
 
 from _test_client_base import RpclibClientTestBase
-from rpclib.client.zeromq import Client
+from rpclib.client.zeromq import ZeroMQClient
 from rpclib.test.interop.server.soap_http_basic import soap_application
 
 class TestRpclibZmqClient(RpclibClientTestBase,unittest.TestCase):
     def setUp(self):
-        self.client = Client('tcp://localhost:5555', soap_application)
+        self.client = ZeroMQClient('tcp://localhost:5555', soap_application)
         self.ns = "rpclib.test.interop.server._service"
 
 if __name__ == '__main__':